update to have lazy loading prop, remove LazyLoadCardContent, move LazyLoadImage to another folder and rename
