+++ src/LazyLoadImage/LazyLoadImage.types.ts
